chore(app-correction) creates midleware for article controller [starts #168927066]
