--- server/helpers/articlehelper.js
@@ -1,1 +1,0 @@
-//to contain validation i will use in articleControllerClass
